pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
authors = [
    {name = "MuhutDil", github = "https://github.com/MuhutDil"}
    ]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"
